requirements.txt
streamlit
pandas
pyyaml
scikit-learn
numpy<2.0
